test/python/requirements.txt
pytest-docker-compose
grpcio-tools
autopep8
urllib
